Module cache::bucket

Leaky bucket algorithm is implemantation for mcatpcha using batch processing Everytime count is increased for an mcaptcha object, a decrement job is added to a batch that is scheduled to be executed at that mcaptcha object’s expiry rate(MCaptcha.get_duration())

Modules

type_methods

Structs

Bucket

Enums

Format

encoding formats for persistence

Constants

REDIS_MCAPTCHA_BUCKET_TYPE_VERSION 🔒

Bucket type version, aka encoding version

Statics

MCAPTCHA_BUCKET_TYPE