Struct cache::Cache

pub struct Cache<K> { /* fields omitted */ }

A thread-safe concurrent cache from key K to arbitrarily typed values

Methods

impl<K: Hash + Eq> Cache<K>

fn new(num_pages: usize, page_size: usize) -> Self

Create a new cache setting page size and numbr of pages.

The page size determines the maximum size of values that can be stored in the cache.

The num pages determines how many slabs of memory of this size should be allocated.

Each page has its own read-write lock, so the more pages you have, the less likely you are to have lock contention.

fn insert<V: 'static + Debug>(&self, key: K, val: V) -> Reference<V>

Insert a value V into cache with key K, returns a Reference to the newly stored or spilled value

fn get<'a, V: 'static + Debug>(&'a self, key: &K) -> Option<Reference<'a, V>>

Is this value in the cache?